# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Patchman-engine API",
        "description": "API of the Patch application on [console.redhat.com](https://console.redhat.com)\n\nSyntax of the `filter[name]` query parameters is described in  [Filters documentation](https://github.com/RedHatInsights/patchman-engine/wiki/API-custom-filters)",
        "contact": {},
        "license": {
            "name": "GPLv3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "v1.18.52"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/patch/v1/packages/{package_name}/systems": {
            "get": {
                "summary": "Show me all my systems which have a package installed",
                "description": "Show me all my systems which have a package installed",
                "operationId": "packageSystems",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit for paging, set -1 to return all",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset for paging",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "package_name",
                        "in": "path",
                        "description": "Package name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Tag filter",
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "filter[system_profile][sap_system]",
                        "in": "query",
                        "description": "Filter only SAP systems",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[system_profile][sap_sids][in]",
                        "in": "query",
                        "description": "Filter systems by their SAP SIDs",
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "filter[system_profile][ansible]",
                        "in": "query",
                        "description": "Filter systems by ansible",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[system_profile][ansible][controller_version]",
                        "in": "query",
                        "description": "Filter systems by ansible version",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[system_profile][mssql]",
                        "in": "query",
                        "description": "Filter systems by mssql version",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[system_profile][mssql][version]",
                        "in": "query",
                        "description": "Filter systems by mssql version",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/controllers.PackageSystemsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ]
            }
        },
...
        "/api/patch/v1/systems/{inventory_id}": {
            "get": {
                "summary": "Show me details about a system by given inventory id",
                "description": "Show me details about a system by given inventory id",
                "operationId": "detailSystem",
                "parameters": [
                    {
                        "name": "inventory_id",
                        "in": "path",
                        "description": "Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/controllers.SystemDetailResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete system by inventory id",
                "description": "Delete system by inventory id",
                "operationId": "deletesystem",
                "parameters": [
                    {
                        "name": "inventory_id",
                        "in": "path",
                        "description": "Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ]
            }
        },
...
        "/api/patch/v1/systems/{inventory_id}/packages": {
            "get": {
                "summary": "Show me details about a system packages by given inventory id",
                "description": "Show me details about a system packages by given inventory id",
                "operationId": "systemPackages",
                "parameters": [
                    {
                        "name": "inventory_id",
                        "in": "path",
                        "description": "Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit for paging, set -1 to return all",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset for paging",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Find matching text",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[description]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[evra]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[summary]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[updatable]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/controllers.SystemPackageResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "controllers.FilterData": {
                "type": "object",
                "properties": {
                    "op": {
                        "type": "string"
                    },
                    "values": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "controllers.Links": {
                "type": "object",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "/link/to/the/first"
                    },
                    "last": {
                        "type": "string",
                        "example": "/link/to/the/last"
                    },
                    "next": {
                        "type": "string",
                        "example": "/link/to/the/next"
                    },
                    "previous": {
                        "type": "string",
                        "example": "/link/to/the/previous"
                    }
                }
            },
            "controllers.ListMeta": {
                "type": "object",
                "properties": {
                    "filter": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/controllers.FilterData"
                        },
                        "description": "Used filters"
                    },
                    "limit": {
                        "type": "integer",
                        "description": "Used response limit (page size) - pagination",
                        "example": 20
                    },
                    "offset": {
                        "type": "integer",
                        "description": "Used response offset - pagination",
                        "example": 0
                    },
                    "search": {
                        "type": "string",
                        "description": "Used search terms",
                        "example": "kernel"
                    },
                    "sort": {
                        "type": "array",
                        "description": "Used sorting fields",
                        "example": [
                            "name"
                        ],
                        "items": {
                            "type": "string"
                        }
                    },
                    "subtotals": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "description": "Some subtotals used by some endpoints"
                    },
                    "total_items": {
                        "type": "integer",
                        "description": "Total items count to return",
                        "example": 1000
                    }
                }
            },
...
            "controllers.PackageSystemItem": {
                "type": "object",
                "properties": {
                    "available_evra": {
                        "type": "string"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "installed_evra": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/controllers.SystemTag"
                        }
                    },
                    "updatable": {
                        "type": "boolean"
                    }
                }
            },
            "controllers.PackageSystemsResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/controllers.PackageSystemItem"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/controllers.Links"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/controllers.ListMeta"
                    }
                }
            },
...
            "controllers.SystemDetailResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/controllers.SystemItem"
                    }
                }
            },
...
            "controllers.SystemItem": {
                "type": "object",
                "properties": {
                    "attributes": {
                        "$ref": "#/components/schemas/controllers.SystemItemAttributes"
                    },
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            },
            "controllers.SystemItemAttributes": {
                "type": "object",
                "properties": {
                    "created": {
                        "type": "string"
                    },
                    "culled_timestamp": {
                        "type": "string"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "insights_id": {
                        "type": "string"
                    },
                    "last_evaluation": {
                        "type": "string"
                    },
                    "last_upload": {
                        "type": "string"
                    },
                    "os": {
                        "type": "string"
                    },
                    "os_major": {
                        "type": "string"
                    },
                    "os_minor": {
                        "type": "string"
                    },
                    "os_name": {
                        "type": "string"
                    },
                    "other_count": {
                        "type": "integer"
                    },
                    "packages_installed": {
                        "type": "integer"
                    },
                    "packages_updatable": {
                        "type": "integer"
                    },
                    "rhba_count": {
                        "type": "integer"
                    },
                    "rhea_count": {
                        "type": "integer"
                    },
                    "rhsa_count": {
                        "type": "integer"
                    },
                    "rhsm": {
                        "type": "string"
                    },
                    "stale": {
                        "type": "boolean"
                    },
                    "stale_timestamp": {
                        "type": "string"
                    },
                    "stale_warning_timestamp": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/controllers.SystemTag"
                        }
                    },
                    "third_party": {
                        "type": "boolean"
                    }
                }
            },
            "controllers.SystemPackageData": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "evra": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "summary": {
                        "type": "string"
                    },
                    "updatable": {
                        "type": "boolean"
                    },
                    "updates": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.PackageUpdate"
                        }
                    }
                }
            },
...
            "controllers.SystemPackageResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/controllers.SystemPackageData"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/controllers.Links"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/controllers.ListMeta"
                    }
                }
            },
            "controllers.SystemTag": {
                "type": "object",
                "properties": {
                    "key": {
                        "type": "string"
                    },
                    "namespace": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    }
                }
            },
...
            "models.PackageUpdate": {
                "type": "object",
                "properties": {
                    "advisory": {
                        "type": "string"
                    },
                    "evra": {
                        "type": "string"
                    }
                }
            },
            "utils.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "RhIdentity": {
                "type": "apiKey",
                "name": "x-rh-identity",
                "in": "header"
            }
        }
    }
}